+++ Sources/KeyboardKitSwiftUI/System/SystemKeyboardStyle.swift
@@ -17,11 +17,46 @@ import CoreGraphics
  */
 public struct SystemKeyboardStyle {
     
-    public static var backspaceWidth: CGFloat = 50.0
-    public static var bottomRowSpacePercentage: CGFloat = 0.5
-    public static var buttonSpacing: CGFloat = 6.0
-    public static var cornerRadius: CGFloat = 4.0
-    public static var frameHeight: CGFloat = 42.0
-    public static var rowSpacing: CGFloat = 12.0
-    public static var shiftWidth: CGFloat = 50.0
+    public init(
+        backspaceWidth: CGFloat,
+        bottomRowSpacePercentage: CGFloat,
+        buttonSpacing: CGFloat,
+        cornerRadius: CGFloat,
+        frameHeight: CGFloat,
+        rowSpacing: CGFloat,
+        shiftWidth: CGFloat) {
+        self.backspaceWidth = backspaceWidth
+        self.bottomRowSpacePercentage = bottomRowSpacePercentage
+        self.buttonSpacing = buttonSpacing
+        self.cornerRadius = cornerRadius
+        self.frameHeight = frameHeight
+        self.rowSpacing = rowSpacing
+        self.shiftWidth = shiftWidth
+    }
+    
+    public let backspaceWidth: CGFloat
+    public let bottomRowSpacePercentage: CGFloat
+    public let buttonSpacing: CGFloat
+    public let cornerRadius: CGFloat
+    public let frameHeight: CGFloat
+    public let rowSpacing: CGFloat
+    public let shiftWidth: CGFloat
+}
+
+public extension SystemKeyboardStyle {
+    
+    /**
+     The standard style that is used if you do not provide a
+     custom one.
+     */
+    static var standard: SystemKeyboardStyle {
+        SystemKeyboardStyle(
+            backspaceWidth: 50.0,
+            bottomRowSpacePercentage: 0.5,
+            buttonSpacing: 6.0,
+            cornerRadius: 4.0,
+            frameHeight: 42.0,
+            rowSpacing: 12.0,
+            shiftWidth: 50.0)
+    }
 }